refactorizacion 1.0
